Apply Cyril Lavier's debian patches set #2:
02-manpages_bad_whatis.dpatch : this patch is a bit strange, but I will try to explain it.

So in manpages, the whatis section needs to be in the following format : "[manpage filename with the extension] \- [a sentence of presentation]".

For example, in the DS2406.3 manpage, the whatis section was "- Dual Addressable Switch with 1kbit Memory", it as to be "DS2406 \- Dual Addressable Switch with 1kbit Memory".

Even if the manpages covers some sensors, the whatis section must contain the filename without extension.

I know, it's a kind of weird rule, but some packages get refused for less than that.